# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: sassaas
host: localhost
basePath: /v1
schemes:
  - http
  - https

paths:
  /sass:
    get:
      x-swagger-router-controller: Sass
      description: Returns CSS by compiling an input Sass template
      operationId: compileSass
      parameters:
        - name: variables
          in: query
          description: 'List of variables to apply. For instance: color1@#123|font@Sans'
          required: false
          type: string
        - name: entrypoints
          in: query
          description: 'List of entrypoints to apply. For instance: index.scss@index.css|custom.scss@custom.css'
          required: false
          type: string
        - name: output
          in: query
          description: 'Type of output. Valid values are: zip or css; default is zip'
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: "file"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sessions/{sessionId}/logo:
    post:
      x-swagger-router-controller: Sessions
      description: Upload a logo
      operationId: uploadLogo
      parameters:
        - name: sessionId
          in: path
          type: string
        - name: logo
          in: formData
          type: file
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResultResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sessions/{sessionId}/variables:
    put:
      x-swagger-router-controller: Sessions
      description: Set variable values
      operationId: setVariables
      parameters:
        - name: sessionId
          in: path
          type: string
        - name: variables
          in: body
          schema:
            type: array
            items:
              $ref: "#/definitions/KeyValue"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResultResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sessions/{sessionId}/css:
    get:
      x-swagger-router-controller: Sessions
      description: Get compiled CSS output
      operationId: getCSSOutput
      parameters:
        - name: sessionId
          in: path
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: "file"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sessions/{sessionId}/zip:
    get:
      x-swagger-router-controller: Sessions
      description: Get compiled Zip output
      operationId: getZipOutput
      parameters:
        - name: sessionId
          in: path
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: "file"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  ResultResponse:
    required:
      - message
    properties:
      message:
        type: string
  KeyValue:
    required:
      - key
      - value
    properties:
      key:
        type: string
      value:
        type: string